@@ -20,12 +20,14 @@
   },
   "devDependencies": {
     "@angular-builders/jest": "^7.1.1",
+    "@angular-devkit/build-angular": "^0.800.1",
     "@angular-devkit/core": "^0.4.9",
     "@angular-devkit/schematics": "^0.4.9",
-    "@angular/cli": "^6.0.0",
-    "@angular/compiler": "6.0.0",
-    "@angular/compiler-cli": "6.0.0",
-    "@angular/language-service": "6.0.0",
+    "@angular/cli": "^8.0.1",
+    "@angular/compiler": "8.0.0",
+    "@angular/compiler-cli": "8.0.0",
+    "@angular/language-service": "8.0.0",
+    "@ngrx/store": "^8.0.0-rc.1",
     "@types/accounting": "^0.4.1",
     "@types/jasmine": "2.5.45",
     "@types/jest": "23.3.10",
@@ -35,14 +37,13 @@
     "@types/tether-drop": "1.4.7",
     "angular-router-loader": "^0.6.0",
     "angular2-template-loader": "^0.6.2",
-    "awesome-typescript-loader": "^3.2.3",
+    "awesome-typescript-loader": "5.2.1",
     "canonical-path": "0.0.2",
     "clean-webpack-plugin": "^0.1.16",
-    "codelyzer": "4.2.1",
+    "codelyzer": "^5.1.0",
     "concurrently": "^3.5.0",
     "copy-webpack-plugin": "^4.0.1",
     "css-loader": "^0.28.11",
-    "extract-text-webpack-plugin": "3.0.0",
     "file-loader": "^0.11.2",
     "html-loader": "^0.5.1",
     "html-webpack-plugin": "^3.2.0",
@@ -60,50 +61,44 @@
     "ts-helpers": "^1.1.2",
     "tslint": "^5.10.0",
     "tslint-loader": "3.6.0",
-    "typescript": "2.7.2",
+    "typescript": "3.4.5",
     "url-loader": "^0.5.9"
   },
   "dependencies": {
-    "@angular-devkit/build-angular": "^0.6.0",
-    "@angular/animations": "6.0.0",
-    "@angular/common": "6.0.0",
-    "@angular/core": "6.0.0",
-    "@angular/forms": "6.0.0",
-    "@angular/http": "6.0.0",
-    "@angular/platform-browser": "6.0.0",
-    "@angular/platform-browser-dynamic": "6.0.0",
-    "@angular/router": "6.0.0",
-    "@clr/angular": "^0.13.9",
-    "@clr/icons": "^0.11.16",
-    "@clr/ui": "^0.11.16",
-    "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^5.2.0",
-    "@ngrx/entity": "^5.2.0",
-    "@ngrx/router-store": "^5.2.0",
-    "@ngrx/store": "^5.2.0",
-    "@ngrx/store-devtools": "^5.2.0",
-    "@webcomponents/custom-elements": "1.0.0",
+    "@angular/animations": "8.0.0",
+    "@angular/common": "8.0.0",
+    "@angular/core": "8.0.0",
+    "@angular/forms": "8.0.0",
+    "@angular/platform-browser": "8.0.0",
+    "@angular/platform-browser-dynamic": "8.0.0",
+    "@angular/router": "8.0.0",
+    "@clr/angular": "^2.0.0-rc.2",
+    "@clr/icons": "^2.0.0-rc.2",
+    "@clr/ui": "^2.0.0-rc.2",
+    "@ngrx/effects": "8.0.0-rc.1",
+    "@ngrx/entity": "8.0.0-rc.1",
+    "@ngrx/router-store": "8.0.0-rc.1",
+    "@ngrx/store-devtools": "8.0.0-rc.1",
+    "@webcomponents/custom-elements": "^1.0.0",
     "accounting": "^0.4.1",
-    "ajv": "^6.4.0",
+    "acorn": "^6.0.0",
+    "ajv": "^6.9.1",
     "babel-polyfill": "^6.26.0",
-    "bootstrap": "4.0.0-alpha.5",
-    "core-js": "2.4.1",
-    "font-awesome": "4.7.0",
-    "jasmine": "2.8.0",
+    "core-js": "^2.4.1",
+    "font-awesome": "^4.7.0",
+    "jasmine": "^2.8.0",
     "lodash": "^4.17.10",
-    "moment": "2.18.1",
-    "ng2-currency-mask": "4.0.2",
-    "ngx-pipes": "^2.1.2",
+    "moment": "^2.18.1",
+    "ng2-currency-mask": "^4.0.2",
     "ngx-toastr": "^8.5.0",
     "numeral": "^2.0.6",
-    "oidc-client": "1.4.1",
-    "protractor": "5.1.2",
+    "oidc-client": "^1.7.1",
+    "protractor": "^5.1.2",
     "redux": "^4.0.0",
-    "rxjs": "6.1.0",
-    "rxjs-compat": "^6.0.0-rc.0",
-    "tether-drop": "1.4.2",
+    "rxjs": "^6.5.2",
+    "tether-drop": "^1.4.2",
     "web-animations-js": "^2.3.1",
-    "zone.js": "0.8.26"
+    "zone.js": "^0.9.1"
   },
   "-vs-binding": {
     "BeforeBuild": [
