@@ -1,3 +1,5 @@
+
+import {retry} from 'rxjs/operators';
 import { Injectable } from '@angular/core';
 import {
     HttpRequest,
@@ -6,13 +8,13 @@ import {
     HttpInterceptor,
     HttpErrorResponse
 } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/retry';
+import { Observable } from 'rxjs';
+
 
 @Injectable()
 export class RetryInterceptor implements HttpInterceptor {
     intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
         // If the call fails, retry until 5 times before throwing an error
-        return next.handle(request).retry(5);
+        return next.handle(request).pipe(retry(5));
     }
 }
