@@ -3,7 +3,6 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { PayeeSelectDropdownComponent } from './payee-select-dropdown.component';
 import { DropdownComponent } from '../../../shared/components';
 import { FormsModule } from '@angular/forms';
-import { NgPipesModule } from 'ngx-pipes';
 
 describe('PayeeSelectDropdownComponent', () => {
   let component: PayeeSelectDropdownComponent;
@@ -12,7 +11,7 @@ describe('PayeeSelectDropdownComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [ PayeeSelectDropdownComponent, DropdownComponent ],
-      imports: [ FormsModule, NgPipesModule ]
+      imports: [ FormsModule ]
     })
     .compileComponents();
   }));
