@@ -3,29 +3,28 @@ import { DatePipe, CommonModule } from "@angular/common";
 import { LoadingComponent, TypeAheadComponent, CurrencyEditComponent, DropdownComponent } from "./components";
 import { CustomDatePipe } from "./pipes/date.pipe";
 import { BudgetCurrencyPipe } from "./pipes/budget.currency.pipe";
-import { ClarityModule, ClrFormsNextModule } from "@clr/angular";
-import { HttpModule } from "@angular/http";
+import { ClarityModule, ClrFormsModule } from "@clr/angular";
 
 import { FormsModule } from "@angular/forms";
-import { NgPipesModule } from "ngx-pipes";
 import { CurrencyMaskModule } from "ng2-currency-mask";
 import { NavComponent } from "./components/nav/nav.component";
+import { PairsPipe } from "./pipes/pairs.pipe";
+import { GroupByPipe } from "./pipes/group.by.pipe";
 
 @NgModule({
     imports: [
         CommonModule,
         ClarityModule,
-        ClrFormsNextModule,
-        FormsModule,
-        HttpModule,
-        NgPipesModule
+        ClrFormsModule,
+        FormsModule
     ],
     exports: [
         CustomDatePipe,
         BudgetCurrencyPipe,
+        PairsPipe,
+        GroupByPipe,
         ClarityModule,
-        ClrFormsNextModule,
-        NgPipesModule,
+        ClrFormsModule,
         LoadingComponent,
         TypeAheadComponent,
         CurrencyEditComponent,
@@ -41,6 +40,8 @@ import { NavComponent } from "./components/nav/nav.component";
     declarations: [
         CustomDatePipe,
         BudgetCurrencyPipe,
+        PairsPipe,
+        GroupByPipe,
         LoadingComponent,
         TypeAheadComponent,
         CurrencyEditComponent,
